Updated WHAT_UI to reflect the changes in folder hierarchy.
(1) Changed the import to reflect the change of path of file WHAT_UI.py
(2) in “WHATPref” class and in method “select_project_dir”, changed the default path to : getcwd() + '/../Projects/Example' to reflect the new folder organization
(3) Icons folder moved to WHAT folder.

These changes were made because for packaging reason. Instead of creating a single file .exe, an uncompressed executable will be created instead. This will make the package larger, but way faster to execute. this will also make the WHAT main folder more clean and less confusing for the end user because all the “system” file will be in one single file.