Made the creation of a new project an explicit independent action of loading an already existing project & UI improvments.
The reasons for this is that by having the loading and creation of projects within the same action, it was confusing about what was exactly done. So, there was the risk to create new files and new folders all over the place when wanting to only load a project. To solve this issue, both action were clearly separated and the creation of a new project was made explicit and resulted in the creation of a .what file. Loading now needs  to pass through a .what file. So it make the action of creating and loading project more explicit and more clear for the user.

(1) Forced the Search4Station window to pop in the center of the main window.
(2) Forced the Main Window to pop up in the center of the screen.
(3) Created a new widget class that consist in a window where the user can enter various information about the project and create a new project. Once a new project is created, a new filed called <project_name>.what is created which contains the information about the project.
(4) Splitted the button Browse in the Menu Bar in two buttons: New Project... and Open Project...
(5) The method select_project_dir in module WHAT has been changed to search for files with the extension *.what instead of folders. Now, it is impossible to load a project without a .what file.