Waterlvl datum option added
(1) Changed the variable name "WLref" to "WLdatum" in "hydroprint.Hydrograph" class.
(2) A lot of code cleaning and commenting.
(3) Changed how the WaterLVL object in handled in the class  hydroprint.Hydrograph(). Instead of having the point of entry of a new  WaterLVL object only from the method generate_hydrograph(), a new method has been created to explicitly push a new WaterLVL object to the class. This methos is named “set_waterLvl”
(4) Remove the need to input the water level time series in the method “best_fit_waterlvl”. It now use the class instance variable automatically instead.
(5) Changed when the method “best_fit_waterlvl” was called in WHAT. It is now done after a check has been done for the layout.
(6) Replaced all the “self.parent.what_pref.load_pref_file()” by “self.parent.check_project()”.
(7) “self.parent.check_project()” does not need to be in a if statement. It can automatically be run at the start of each method that requires it.
(8) merge the method “check_if_layout_exist” in “load_waterlvl”
(9) Code cleaning of method “load_waterlvl” in WHAT.
(10) method “generate_hydrograph” does not requires explicitly a waterlvlobj as input. It now used the class variable that must now be explicitly set with its own function.