cleanup of class *NewProject* and its use in *WHAT.py*:
(1) Used directly *close* built-in method as slot of button *cancel*.
(2) renamed method *clear_UI* to *reset_UI*
(3) removed the software version argument. Now directly fetched from *db* module.
(4) Dialog window is now destroyed on close.
(5) Reimplemented the *show* method in the class *NewProject*, which center the window to the parent or to the screen if there is no parent.