Production of a Windows Executable: (1) In Windows, still had problem to run the gap-filling routine in batch mode if there was no while loop added to assure that the Thread is closed before processing with the next station. This was added to the `gap_fill_start` method of the `GapFillWeatherGUI` class of the `gapfill_weather_gui` module.
Problem due to the use of the `statsmodels` module:
(2) Had to add `from multiprocessing import freeze_support` and the call `freeze_support()` to the WHAT module. This was causing problem with the executable. See: http://stackoverflow.com/questions/7866312/importerror-no-module-named-multiprocessing.
(3) There was some hook problem with PyInstaller that were resolved by adding 2 silent import, namely: --hidden-import:scipy.linalg.cython_blas and --hidden-import:scipy.linalg.cython_lapack. This was added to the "bat" file used for calling PyInstaller.
(4) There was problem also incorporating Panda. Had to manually take the file "lib.pyd" from the pandas installation folder and move it to the WHAT folder, after renaming it "pandas_lib.pyd". I've read that this is a bug that has been resolved in newer version of PyInstaller.
(5) Had to change the way statmodels was imported in the code. Was causing problem with PyInstaller.