Improvements of Hydrograph plotting:
(1) Bug : corrected an error in the ylabel for Precipitations in french.

(2) Reworked the way the graph title was displayed and worked. Now, it contains the name of the well and of the weather station and is displayed at the top left corner of the graph. It is possible to enable or disable the legend and title in the graph from the *page setup* menu.

(3) Added the legend to the hydrograph in the top right corner.

(4) Added the option to activate the trend on the water level. A moving average is then done on the water level. The trend is shown as a continuous line, while each data points is shown in light blue.

Added a color class to manage the colors in the module *hydrograph*. Colors are now saved and loaded in a file named *colors.db*. It is then possible for the users to change the colors for the graph.