Made Pan/Zoom more intuitive
Removed the button to activate/deactivate pan/zoom. Instead, the
pan/zoom mode is automatically activated when clicking on the graph and
holding either the left or right mouse button. When the mode to add or
delete peak for the MRC or to select a region for the brf is activated,
the pan/zoom option is not activated. This removed a lot of mouse
clicking, while preserving the intended workflow.